requirements.txt
streamlit
openai
plotly
python-dotenv
